NEW FILE IMPLEMENTATION!
Andrew and I created a new file after finding an example of how to implement the Gtk.UIManager. It includes the WORKING CODE for the separate user interface where the SRS settings that MouseTrap user's can change will be located. We just need to work it in with the YAML file and solve the issues with multithreading.